refactor: separate models and repositories per LAD architecture
- Created repositories/ folder with SQL-only database operations
  - CampaignRepository.js
  - CampaignLeadRepository.js
  - CampaignStepRepository.js
  - CampaignLeadActivityRepository.js

- Refactored models/ to be SQL-free (data shapes, validation, mapping only)
  - CampaignModel.js
  - CampaignLeadModel.js
  - CampaignStepModel.js
  - CampaignLeadActivityModel.js

- Updated all controllers to use repositories for DB operations and models for mapping
- Updated services to use repositories
- Cleaned up constants.js to remove duplicated model-specific constants
- All changes follow LAD architecture guidelines for separation of concerns
