+++ backend/features/campaigns/controllers/CampaignLeadsController.js
@@ -4,8 +4,10 @@
  * LAD Architecture Compliant - No SQL in controllers, uses logger instead of console
  */
 
+const CampaignLeadRepository = require('../repositories/CampaignLeadRepository');
 const CampaignLeadModel = require('../models/CampaignLeadModel');
 const { getSchema } = require('../../../core/utils/schemaHelper');
+const CampaignLeadActivityRepository = require('../repositories/CampaignLeadActivityRepository');
 const CampaignLeadActivityModel = require('../models/CampaignLeadActivityModel');
 const logger = require('../../../core/utils/logger');
 
@@ -19,38 +21,20 @@ class CampaignLeadsController {
       const tenantId = req.user.tenantId;
       const { id } = req.params;
       const { status, limit, offset } = req.query;
-      // LAD Architecture: Use model layer instead of direct SQL in controller
-      const schema = getSchema(req);
-      const leads = await CampaignLeadModel.getByCampaignId(id, tenantId, {
+      // LAD Architecture: Use repository layer instead of direct SQL in controller
+      const dbLeads = await CampaignLeadRepository.getByCampaignId(id, tenantId, {
         status,
         limit: parseInt(limit) || 100,
         offset: parseInt(offset) || 0
       }, req);
+      const leads = dbLeads.map(lead => CampaignLeadModel.mapLeadFromDB(lead));
       
       // Format leads for frontend
-      const formattedLeads = leads.map(row => {
+      const formattedLeads = leads.map(lead => {
         try {
-          // Parse JSONB fields safely
-          let snapshot = {};
-          let leadData = {};
-          
-          try {
-            snapshot = typeof row.snapshot === 'string' 
-              ? JSON.parse(row.snapshot || '{}') 
-              : (row.snapshot || {});
-          } catch (e) {
-            logger.warn('[Campaign Leads] Error parsing snapshot', { error: e.message });
-            snapshot = {};
-          }
-          
-          try {
-            leadData = typeof row.lead_data === 'string' 
-              ? JSON.parse(row.lead_data || '{}') 
-              : (row.lead_data || {});
-          } catch (e) {
-            logger.warn('[Campaign Leads] Error parsing lead_data', { error: e.message });
-            leadData = {};
-          }
+          // Use mapped lead data
+          const snapshot = lead.snapshot || {};
+          const leadData = lead.leadData || {};
           
           // Extract profile summary from lead_data if it exists
           const profileSummary = leadData.profile_summary || null;
@@ -58,9 +42,9 @@ class CampaignLeadsController {
           // Extract apollo_person_id from lead_data (needed for reveal email/phone)
           const apolloPersonId = leadData.apollo_person_id || leadData.id || leadData.apollo_id || null;
           
-          // Extract name fields - Priority: snapshot > lead_data
-          const firstName = snapshot.first_name || leadData.first_name || leadData.employee_name?.split(' ')[0] || '';
-          const lastName = snapshot.last_name || leadData.last_name || leadData.employee_name?.split(' ').slice(1).join(' ') || '';
+          // Extract name fields - Priority: mapped fields > snapshot > lead_data
+          const firstName = lead.firstName || snapshot.first_name || leadData.first_name || leadData.employee_name?.split(' ')[0] || '';
+          const lastName = lead.lastName || snapshot.last_name || leadData.last_name || leadData.employee_name?.split(' ').slice(1).join(' ') || '';
           
           // Build full name
           let name = '';
@@ -75,32 +59,32 @@ class CampaignLeadsController {
           }
           
           return {
-            id: row.id,
-            campaign_id: row.campaign_id,
-            lead_id: row.lead_id,
-            status: row.status,
+            id: lead.id,
+            campaign_id: lead.campaign_id,
+            lead_id: lead.lead_id,
+            status: lead.status,
             name: name,
             first_name: firstName,
             last_name: lastName,
-            email: snapshot.email || leadData.email || leadData.employee_email || leadData.work_email || null,
-            phone: snapshot.phone || leadData.phone || leadData.employee_phone || leadData.phone_number || null,
-            company: snapshot.company_name || leadData.company_name || leadData.company || leadData.employee_company || leadData.organization?.name || null,
-            title: snapshot.title || leadData.title || leadData.employee_title || leadData.job_title || leadData.headline || null,
-            linkedin_url: snapshot.linkedin_url || leadData.linkedin_url || leadData.employee_linkedin_url || leadData.linkedin || null,
+            email: lead.email || snapshot.email || leadData.email || leadData.employee_email || leadData.work_email || null,
+            phone: lead.phone || snapshot.phone || leadData.phone || leadData.employee_phone || leadData.phone_number || null,
+            company: lead.companyName || snapshot.company_name || leadData.company_name || leadData.company || leadData.employee_company || leadData.organization?.name || null,
+            title: lead.title || snapshot.title || leadData.title || leadData.employee_title || leadData.job_title || leadData.headline || null,
+            linkedin_url: lead.linkedinUrl || snapshot.linkedin_url || leadData.linkedin_url || leadData.employee_linkedin_url || leadData.linkedin || null,
             photo_url: leadData.photo_url || leadData.employee_photo_url || leadData.avatar || snapshot.photo_url || null,
             profile_summary: profileSummary,
             apollo_person_id: apolloPersonId, // Include apollo_person_id for reveal functionality
-            created_at: row.created_at,
-            updated_at: row.updated_at
+            created_at: lead.created_at,
+            updated_at: lead.updated_at
           };
         } catch (formatError) {
           logger.error('[Campaign Leads] Error formatting lead', { error: formatError.message, stack: formatError.stack });
           // Return minimal data if formatting fails
           return {
-            id: row.id,
-            campaign_id: row.campaign_id,
-            lead_id: row.lead_id,
-            status: row.status,
+            id: lead.id,
+            campaign_id: lead.campaign_id,
+            lead_id: lead.lead_id,
+            status: lead.status,
             name: 'Unknown',
             first_name: null,
             last_name: null,
@@ -111,8 +95,8 @@ class CampaignLeadsController {
             linkedin_url: null,
             photo_url: null,
             profile_summary: null,
-            created_at: row.created_at,
-            updated_at: row.updated_at
+            created_at: lead.created_at,
+            updated_at: lead.updated_at
           };
         }
       });
@@ -148,7 +132,8 @@ class CampaignLeadsController {
         });
       }
 
-      const createdLeads = await CampaignLeadModel.bulkCreate(id, tenantId, leads);
+      const dbLeads = await CampaignLeadRepository.bulkCreate(id, tenantId, leads, req);
+      const createdLeads = dbLeads.map(lead => CampaignLeadModel.mapLeadFromDB(lead));
 
       res.status(201).json({
         success: true,
@@ -174,12 +159,13 @@ class CampaignLeadsController {
       const { id } = req.params;
       const { status, stepType, limit, offset } = req.query;
 
-      const activities = await CampaignLeadActivityModel.getByCampaignId(id, tenantId, {
+      const dbActivities = await CampaignLeadActivityRepository.getByCampaignId(id, tenantId, {
         status,
         stepType,
         limit: parseInt(limit) || 1000,
         offset: parseInt(offset) || 0
-      });
+      }, req);
+      const activities = dbActivities.map(activity => CampaignLeadActivityModel.mapActivityFromDB(activity));
 
       res.json({
         success: true,
@@ -194,7 +180,7 @@ class CampaignLeadsController {
       });
     }
   }
-
+}
 
 module.exports = CampaignLeadsController;
